+++ spec/Runner.html
@@ -16,6 +16,7 @@
   <script src="../src/rpd.js"></script>
 
   <!-- include spec files here... -->
+  <script src="./matchers.js"></script>
   <script src="./building.spec.js"></script>
   <script src="./events.spec.js"></script>
   <script src="./registration.spec.js"></script>
